رفع اشکال

+++ ApiKarbord/Models/Public/Web_IDocR.cs
@@ -274,7 +274,7 @@
         [StringLength(50)]
         public string MainUnitName { get; set; }
 
-        public int? DeghatR { get; set; }
+        public byte? DeghatR { get; set; }
 
         [Key]
         [Column(Order = 12)]
